backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
pydantic-settings
python-telegram-bot
httpx
redis
pydantic[email]
sqlalchemy
psycopg
psycopg[binary]==3.2.10
pytz==2025.2
python-dotenv==1.1.1
asyncpg